This is the clean working project
I checked at least 10 times it runs the only thing is that the color for the
main activity is blue not sure why. But I will fix now lets see how much we can finish
in a day. We should work on the main activity and profile at the same time